fabsf instead of fabs because RSGL is just that fab

+++ RSGL.h
@@ -1062,7 +1062,7 @@ void RSGL_drawTriangleF(RSGL_triangleF t, RSGL_color c) {
 
 void RSGL_drawTriangleHyp(RSGL_pointF p, size_t angle, float hypotenuse, RSGL_color color) {
     float dir = (hypotenuse > 0);
-    hypotenuse = fabs(hypotenuse);
+    hypotenuse = fabsf(hypotenuse);
 
     float base = hypotenuse * (cos(angle) * RAD2DEG);
     float opp = hypotenuse * (cos(angle) * RAD2DEG); 
@@ -1190,7 +1190,7 @@ void RSGL_drawTriangleFOutline(RSGL_triangleF t, u32 thickness, RSGL_color c) {
     RSGL_point3DF points[] = {{t.p3.x, t.p3.y, 0.0f}, {t.p1.x, t.p1.y, 0.0f}, {t.p1.x, t.p1.y, 0.0f}, {t.p2.x, t.p2.y, 0.0f}, {t.p2.x, t.p2.y, 0.0f}, {t.p3.x, t.p3.y, 0.0f}};
     RSGL_point3DF texPoints[] = {{0, 0, 0.0f}, {0, 0, 0.0f}, {0, 0, 0.0f}, {0, 0, 0.0f}, {0, 0, 0.0f}, {0, 0, 0.0f}};
 
-    RSGL_rectF r = {t.p2.x, t.p3.y, fabs(t.p2.x - t.p1.x), fabs(t.p2.y - t.p3.y)};
+    RSGL_rectF r = {t.p2.x, t.p3.y, fabsf(t.p2.x - t.p1.x), fabsf(t.p2.y - t.p3.y)};
 
     RSGL_BASIC_DRAW(RGL_LINES, (RSGL_point3DF*)points, (RSGL_point3DF*)texPoints, r, c, 6);
     rglLineWidth(1);
